[FIX] Fix pull request from Johan
