+++ .idea/modules.xml
@@ -3,6 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/Comics_Project.iml" filepath="$PROJECT_DIR$/Comics_Project.iml" />
+      <module fileurl="file://$PROJECT_DIR$/Comics_Project.iml" filepath="$PROJECT_DIR$/Comics_Project.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
