--- app/src/main/java/nicolas/quillon/iem/gestion_comics/Modele/pojo/Creators.java
@@ -1,10 +1,0 @@
-package nicolas.quillon.iem.gestion_comics.Modele.pojo;
-
-import java.util.ArrayList;
-
-/**
- * Created by iem on 08/12/2017.
- */
-
-public class Creators extends ArrayList<Creator>{
-}
