+++ app/src/main/java/nicolas/quillon/iem/gestion_comics/ui/Vue/MainActivity.java
@@ -1,13 +1,12 @@
 package nicolas.quillon.iem.gestion_comics.ui.Vue;
 
+import android.os.Build;
+import android.support.annotation.RequiresApi;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.widget.ImageView;
 import android.widget.ListView;
-import android.widget.TextView;
 
-import org.w3c.dom.Text;
-
+import nicolas.quillon.iem.gestion_comics.Modele.manager.JSONManager;
 import nicolas.quillon.iem.gestion_comics.Modele.pojo.Comics;
 import nicolas.quillon.iem.gestion_comics.R;
 
@@ -16,16 +15,22 @@ public class MainActivity extends AppCompatActivity {
     //region Variables
     private ListView listViewComic;
     private Comics listComics;
+    private JSONManager jsonManager;
 
     private ListAdapterComics adapterComics;
 
     //endregion
 
     //region Override
+    @RequiresApi(api = Build.VERSION_CODES.KITKAT)
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        //Initialisation du fichier JSON
+        jsonManager=new JSONManager("/data/sample-ok.json"); //dossier data à la racine du tel (pas sur SD)
+        jsonManager.init();
 
         //ad - appel de la fonction d'initialisation
         InitializeView();
@@ -43,7 +48,7 @@ public class MainActivity extends AppCompatActivity {
         listViewComic = (ListView) findViewById(R.id.listViewComics);
         listComics = new Comics();
 
-        adapterComics = new ListAdapterComics(this, listComics);
+        adapterComics = new ListAdapterComics(this, jsonManager.getAll().getResults());
 
         listViewComic.setAdapter(adapterComics);
